After bisecting for first bad commit, get commit date.
We use Porting/bisect.pl -- a wrapper around 'git bisect' -- to identify
problematic commits within the Perl 5 core distribution.  This program
can be used when our objective is to get the "first bad commit" -- i.e.,
the SHA of the commit which "broke" the code.  The program will output
the first five lines of the output from 'git log --format=fuller' for
the "bad" commit, thereby displaying the committer and the commit date.
('git bisect' only displays the author and the patch date, but the
committer and commit date are often useful as well.)